% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{installIfNeeded}
\alias{installIfNeeded}
\title{Install packages or packages from a CRAN repository if not already installed}
\usage{
installIfNeeded(packages, ...)
}
\arguments{
\item{packages}{a character vector (one or more) of packages to
install. Dependencies will be installed as well.}

\item{...}{other args passed to \code{\link{install.packages}}}
}
\description{
Install a package or packages if it is not already
    installed. By default, the repository used is the cloud
    repository. This has no effect if the packages are already
    installed and is, therefore, a safe way to ensure necessary
    packages are available to be used in subsequent code. So in two
    invocations of this function with the same arguments, only the
    first may have an effect. See examples.
}
\examples{
\dontrun{
installIfNeeded("survival")  ## has no effect
installIfNeeded("dplyr") ## will install the dependencies as well if not installed.
}

}
