% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stats101_setup}
\alias{stats101_setup}
\title{stats101_setup}
\usage{
stats101_setup()
}
\description{
Set up the R environment by installing all required
    packages for the course. CRAN packages are installed as binary,
    even if a newer source package is available. Github packages
    are installed from source, but they better not assume
    availability of developer tools on the target machine.
}
\examples{
\dontrun{
installIfNeeded("survival")  ## has no effect
installIfNeeded("dplyr") ## will install the dependencies as well if not installed.
}
}
